Provide consistent exception handling.
All exceptions provided by attribute mapper will now be derived from
`AttributeMapperException`.

Also provide a new JSON Schema to validate JSON/YAML policies in order
to get meaningful error messages.

The following are the description of new exceptions:

AttributeMapperException extends Exception:

Base class for all other exceptions. This is also used to wrap
exceptions we don't understand. In this case, the message will be
generic : "Error while processing mapping policy".  This may be an
internal issue with the attribute mapper.

KnownAttributeMapperException extends AttributeMapperException:

Base class for all "known" exceptions.  In this case the message is
specific to the exception and may be sent to the end user.

ParseException extends KnownAttributeMapperException:

Malformed XML, JSON, YAML.

ValidationException extends KnownAttributeMapperException:

There's a validation error -- something which validates XSD or
JsonSchema.

XPathException extends ValidationException

There's an issue with an XPath in the policy.